packages/backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1

# Authentication
PyJWT[crypto]>=2.8.0
httpx>=0.26.0
authlib>=1.3.0

# Utilities
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Templating & Admin GUI
jinja2>=3.1.0
python-slugify>=8.0.0

# Security
cryptography>=42.0.0
